Make ServiceImpl test cases hermetic
Ensures that adding one package to our monolithic list doesn't require
that we go back and modify N tests for a given change.  This does the
lazy thing of repeating all the packages we previously had in PopulateDB
across all tests without caring for the specifics of which tests
"needed" which packages exactly.